+++ lib/vanagon/utilities/extra_files_signer.rb
@@ -19,11 +19,13 @@ class Vanagon
             remote_host = "#{project.signing_username}@#{project.signing_hostname}"
             remote_destination_path = "#{remote_host}:#{tempdir}"
             remote_file_location = "#{remote_host}:#{file_location}"
+            extra_flags = ''
+            extra_flags = '--extended-attributes' if project.platform.is_macos?
 
             commands += [
-              "rsync -e '#{Vanagon::Utilities.ssh_command}' --verbose --recursive --hard-links --links  --no-perms --no-owner --no-group #{local_source_path} #{remote_destination_path}",
+              "rsync -e '#{Vanagon::Utilities.ssh_command}' --verbose --recursive --hard-links --links  --no-perms --no-owner --no-group #{extra_flags} #{local_source_path} #{remote_destination_path}",
               "#{Vanagon::Utilities.ssh_command} #{remote_host} #{project.signing_command} #{file_location}",
-              "rsync -e '#{Vanagon::Utilities.ssh_command}' --verbose --recursive --hard-links --links  --no-perms --no-owner --no-group #{remote_file_location} #{local_source_path}"
+              "rsync -e '#{Vanagon::Utilities.ssh_command}' --verbose --recursive --hard-links --links  --no-perms --no-owner --no-group #{extra_flags} #{remote_file_location} #{local_source_path}"
             ]
           end